Move useTitleupdater into composables
